@@ -1,11 +1,11 @@
 name: future_alert
 description: Flutter package to display alerts
-version: 0.0.4
-author: Tomas Garay <tomasgaraypaz@gmail.com>, Nelson Agurcial<jagurcia21@hotmail.com>
+version: 0.0.5
+author: Nelson Agurcial<jagurcia21@hotmail.com>
 homepage: https://gitlab.com/Josg182/future_alerts.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
