@@ -51,50 +51,60 @@ class _MyHomePageState extends State<MyHomePage> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            RaisedButton(
-              color: Colors.green[800],
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(8.0)
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.green[800],
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(8.0)
+                )
               ),
               child: Text("Done", style: TextStyle(color: Colors.white),),
               onPressed: (){
                 FutureAlert.done(context, "Done", "This alert has shown correctly");
               },
             ),
-            RaisedButton(
-              color: Colors.red[800],
-              shape: RoundedRectangleBorder(
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.red[800],
+                shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(8.0)
+                )
               ),
               child: Text("Error", style: TextStyle(color: Colors.white),),
               onPressed: (){
                 FutureAlert.error(context, "Error", "You shouldn't have done that");
               },
             ),
-            RaisedButton(
-              color: Colors.yellow[800],
-              shape: RoundedRectangleBorder(
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.yellow[800],
+                shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(8.0)
+                )
               ),
               child: Text("Warning", style: TextStyle(color: Colors.white),),
               onPressed: (){
                 FutureAlert.warning(context, "Warning!", "Wow that didn't go well");
               },
             ),
-            RaisedButton(
-              color: Colors.blue[800],
-              shape: RoundedRectangleBorder(
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.blue[800],
+                shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(8.0)
+                )
               ),
               child: Text("Info", style: TextStyle(color: Colors.white),),
               onPressed: (){
                 FutureAlert.info(context, "Info", "Hey!, I just wanted to say hello");
               },
             ),
-            RaisedButton(
-              color: Colors.blue[800],
-              shape: RoundedRectangleBorder(
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.blue[800],
+                shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(8.0)
+                )
               ),
               child: Text("Question", style: TextStyle(color: Colors.white),),
               onPressed: (){
@@ -103,10 +113,12 @@ class _MyHomePageState extends State<MyHomePage> {
                 });
               },
             ),
-            RaisedButton(
-              color: Colors.orange[800],
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(8.0)
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.orange[800],
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(8.0)
+                )
               ),
               child: Text("Custom", style: TextStyle(color: Colors.white),),
               onPressed: (){
@@ -121,6 +133,7 @@ class _MyHomePageState extends State<MyHomePage> {
                       color: Colors.blue[800],
                       onPressed: (){
                         print(true);
+                        Navigator.of(context).pop(true);
                       },
                     )
                   ]
